Added state logic to turn valve on and off
The valve can be turned on and off. The valve is set to turn on for 60 seconds and then turn off for the time set in the valve settings display0 state. The display2 state indicated if the valve is open or close and for how long it has been and will be.