Open and Close Valve
Pneumatic valve opens and close based on user defined close duration and the default 1 min open duration. The valve controls the water flow for the misting system. The water reservoir is constantly pressurized. The valve is open to release the pressurized mist. The valve is closed to stop the mist.