Start/Stop Data Logging
Multiple data files can be stored to the SD card. Hold down the button while on the "enviro data" display to start and stop data logging. When a new data log is started a new file is generated. The file name can be specified in code; default is "DATA". The file type can be specified as well in code; deffault is ".txt". The new data files follow the naming scheme Data0.txt, Data1.txt, Data2.txt, ...